virtualenv_requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
Flask-WTF==0.9.5
